requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Lightning
lightning>=2.0.0

# Config
omegaconf>=2.3.0

# Data
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0

# ML utilities
scikit-learn>=1.2.0
iterstrat>=0.1.0

# Logging and monitoring
wandb>=0.15.0

# Progress bars
rich>=13.0.0
tqdm>=4.65.0
